requirements.txt
# Challenge Toolkit v2 - Dependencies

# Core data processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Machine Learning
scikit-learn>=1.3.0
umap-learn>=0.5.0

# Visualization
plotly>=5.18.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web application
streamlit>=1.28.0

# Configuration
pyyaml>=6.0

# Optional: For enhanced performance
# numba>=0.58.0
# joblib>=1.3.0
